Entering in the last stage of the first release refactoring....

+++ VehicleManager/src/main/java/com/pt/pires/services/integrator/VehicleIntegratorService.java
@@ -22,38 +22,68 @@ public class VehicleIntegratorService implements IVehicleIntegratorService{
 	@Qualifier("fileService")
 	private IFileService fileService;
 	
+	
+	public VehicleIntegratorService(IVehicleService vs,IFileService fs) {
+		this.localVehicleService = vs;
+		this.fileService = fs;
+	}
+	
+	public VehicleIntegratorService() { }
+	
 	@Override
-	public void createUnlicensedVehicle(String vehicleName, String brand, Date acquisitionDate,byte[] image)
-			throws VehicleManagerException {
+	public void createUnlicensedVehicle(String vehicleName, String brand, Date acquisitionDate,
+			boolean image, byte[] imageFile)throws VehicleManagerException {
+		if(vehicleName == null || brand == null || acquisitionDate == null){
+			throw new IllegalArgumentException();
+		}
 		localVehicleService.createUnlicensedVehicle(vehicleName, brand, acquisitionDate);
-		try{
-			fileService.addPortraitImage(vehicleName, image);
-		}catch(VehicleManagerException e){
-			localVehicleService.removeVehicle(vehicleName);
-			throw new ImpossibleSaveFileException();
+		if(image){
+			try{
+				fileService.addPortraitImage(vehicleName, imageFile);
+			}catch(VehicleManagerException e){
+				localVehicleService.removeVehicle(vehicleName);
+				throw new ImpossibleSaveFileException();
+			}
 		}
 	}
 
 	@Override
 	public void createLicensedVehicle(String vehicleName, String brand, Date acquisitionDate,
-			String license, Date licenseDate, byte[] image) throws VehicleManagerException {
+			String license, Date licenseDate,boolean image, byte[] imageFile) throws VehicleManagerException {
+		if(vehicleName == null || brand == null || acquisitionDate == null || licenseDate == null
+				|| license == null){
+			throw new IllegalArgumentException();
+		}
 		localVehicleService.createLicensedVehicle(vehicleName, brand, acquisitionDate, license, licenseDate);
-		try{
-			fileService.addPortraitImage(vehicleName, image);
-		}catch(VehicleManagerException e){
-			localVehicleService.removeVehicle(vehicleName);
-			throw new ImpossibleSaveFileException();
+		if(image){
+			try{
+				fileService.addPortraitImage(vehicleName, imageFile);
+			}catch(VehicleManagerException e){
+				localVehicleService.removeVehicle(vehicleName);
+				throw new ImpossibleSaveFileException();
+			}
 		}
 	}
 	
 	@Override
 	public void removeVehicle(String vehicleName) throws VehicleManagerException {
+		if(vehicleName == null) {
+			throw new IllegalArgumentException();
+		}
 		localVehicleService.removeVehicle(vehicleName);
 		try {
 			fileService.removeVehicleFiles(vehicleName);
 		} catch (VehicleManagerException e) {
 			e.printStackTrace();
 		}
+	}
+
+	@Override
+	public void addOrUpdateVehiclePortrait(String vehicleName, byte[] imageFile) throws VehicleManagerException {
+		if(vehicleName == null || imageFile == null) {
+			throw new IllegalArgumentException();
+		}
+		fileService.addPortraitImage(vehicleName, imageFile);
 	}
 
 }
